Add files via upload
The Sean Keith Smith conjecture is not a proven theorem. It states a connection between the last digit of an even perfect number and the primality of the corresponding Mersenne number (2^p - 1).
While the conjecture holds true for many known Mersenne primes, it hasn't been mathematically proven for all cases. There could be exceptions.